Refactor unit tests and karma configuration
This commit refactors the unit tests for the app component and updates the karma configuration. The unit tests for the app component have been rewritten using the TestBed and async functions from the '@angular/core/testing' module. The tests now check if the app component is created, if it has the correct title, and if the title is rendered correctly in the HTML. The karma configuration file (karma.conf.js) has been added to the repository and includes the necessary plugins and preprocessors for running the unit tests. The webpack configuration in the karma file has been updated to use the 'ts-loader' for compiling TypeScript files. The commit also includes minor formatting changes to the karma.conf.js file.